01_crawl/requirements.txt
httpx
pypdf 
python-dotenv

chromadb
opentelemetry-api==1.39.1
opentelemetry-exporter-gcp-logging==1.10.0a0
opentelemetry-exporter-otlp-proto-http==1.39.1
langchain-community
langchain-google-genai
langchain-chroma

google-adk==1.14.1
google-cloud-aiplatform==1.126.1
google-cloud-discoveryengine
